pyproject.toml
[project]
name = "taska-v2"
version = "0.1.0"
description = "Indian Cultural Dataset Pipeline"
authors = [
    { name = "Prince" }
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "requests==2.32.3",
    "Pillow==10.4.0",
    "numpy==1.26.4",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "python-dotenv==1.0.1",
    "beautifulsoup4==4.12.3",
    "lxml==5.2.2",
]

[tool.uv]
# uv specific settings (optional)

[tool.setuptools]
packages = ["pipeline", "sources", "utils"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"